Add support for Java 8 lambdas + test
Java 8 lambdas generate synthetic methods used in invokedynamic
instructions. The synthetic methods need to be copied and renamed,
similarly to access methods. Bootstrap method entries need to be
adjusted; in ASM, this is done through the visitInvokeDynamicInsn()
method directly as the ASM class writer takes care of the constant
pool and bootstrap method table for us.

git-svn-id: https://cofoja.googlecode.com/svn/trunk@156 54da9a4b-e3f5-3583-49f8-1ce3445e53fc